Add CSV splitting functionality
Implemented the main functionality to split a CSV file into smaller CSV files. This newly added command-line tool, written in Go, can take a CSV file as an argument, splitting it into various CSV files, each with a maximum of 150 rows. The files are created and saved sequentially. Also, a detailed usage and description of this feature have been added to the readme file.